refactor: controller와 내부 애플리케이션에서 동일하게 사용했던 request 를 controller만 사용하고, 내부에는 command 객체로 데이터를 받도록 분리
