+++ src/main/java/org/wedding/adapter/in/web/CardController.java
@@ -11,8 +11,9 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 import org.wedding.adapter.in.web.dto.CreateCardRequest;
 import org.wedding.adapter.in.web.dto.ModifyCardRequest;
-import org.wedding.application.port.in.usecase.CreateCardUseCase;
-import org.wedding.application.port.in.usecase.ModifyCardUseCase;
+import org.wedding.application.port.in.command.card.CreateCardCommand;
+import org.wedding.application.port.in.usecase.card.CreateCardUseCase;
+import org.wedding.application.port.in.usecase.card.ModifyCardUseCase;
 import org.wedding.common.response.ApiResponse;
 
 import io.swagger.v3.oas.annotations.Operation;
@@ -35,7 +36,11 @@ public class CardController {
     @Operation(summary = "카드 생성", description = "카드 생성")
     @ResponseStatus(HttpStatus.CREATED)
     public ResponseEntity<ApiResponse<Void>> createCard(@RequestBody @Valid CreateCardRequest request) {
-        createCardUseCase.createCard(request);
+
+        CreateCardCommand command = new CreateCardCommand(request.cardTitle(), request.budget(), request.deadline());
+
+        createCardUseCase.createCard(command);
+
         ApiResponse<Void> response = ApiResponse.successApiResponse(HttpStatus.CREATED, "카드가 생성되었습니다.");
         return new ResponseEntity<>(response, response.status());
     }
