--- src/main/java/org/wedding/application/port/in/usecase/CreateCardUseCase.java
@@ -1,8 +1,0 @@
-package org.wedding.application.port.in.usecase;
-
-import org.wedding.adapter.in.web.dto.CreateCardRequest;
-
-public interface CreateCardUseCase {
-
-    void createCard(CreateCardRequest request);
-}
